requirements.txt
django==2.2.17
djangorestframework==3.8.2
djangorestframework-jwt
Pillow
psycopg2-binary
django-simple-email-confirmation
django-cors-headers
dj-database-url
python-decouple
